INS-1414: pulse tracker split
